# Faster Kohel isogenies without bivariate polynomials